refactor: 중복 기능 부분 분리
