+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,12 +1,9 @@
 package chess.domain.piece;
 
-import chess.domain.board.Board;
-import chess.domain.board.Horizontal;
 import chess.domain.board.Position;
-import chess.domain.board.Vertical;
+import chess.domain.piece.moveStrategy.BishopMovement;
 import chess.domain.piece.moveStrategy.ToEndOfLineStrategy;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class Bishop extends Piece {
@@ -19,37 +16,7 @@ public class Bishop extends Piece {
 
     @Override
     public List<Position> searchMovablePositions(Position target) {
-        List<Position> movablePositions = new ArrayList<>();
-
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.RIGHT_TOP));
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.LEFT_TOP));
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.LEFT_BOTTOM));
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.RIGHT_BOTTOM));
-
-        return movablePositions;
-    }
-
-    private List<Position> calculateBishopMovablePositions(Position target, Direction direction) {
-        List<Position> result = new ArrayList<>();
-        int horizontalWeight = target.getHorizontalWeight();
-        int verticalWeight = target.getVerticalWeight();
-        while (isInsideBoard(target, direction)) {
-            horizontalWeight += direction.getX();
-            verticalWeight += direction.getY();
-            result.add(
-                    Position.of(Horizontal.findFromWeight(horizontalWeight), Vertical.findFromWeight(verticalWeight))
-            );
-        }
-        return result;
-    }
-
-    private boolean isInsideBoard(Position target, Direction direction) {
-        int horizontalWeight = target.getHorizontalWeight();
-        int verticalWeight = target.getVerticalWeight();
-
-        return horizontalWeight + direction.getX() >= Board.MIN_BORDER
-                && horizontalWeight + direction.getX() <= Board.MAX_BORDER
-                && verticalWeight + direction.getY() >= Board.MIN_BORDER
-                && verticalWeight + direction.getY() <= Board.MAX_BORDER;
+        BishopMovement bishopMovement = new BishopMovement();
+        return bishopMovement.searchMovablePositions(target);
     }
 }
