+++ src/main/java/chess/domain/piece/King.java
@@ -1,12 +1,9 @@
 package chess.domain.piece;
 
-import chess.domain.board.Board;
-import chess.domain.board.Horizontal;
 import chess.domain.board.Position;
-import chess.domain.board.Vertical;
+import chess.domain.piece.moveStrategy.KingMovement;
 import chess.domain.piece.moveStrategy.SpecifiedLocationStrategy;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class King extends Piece {
@@ -19,33 +16,7 @@ public class King extends Piece {
 
     @Override
     public List<Position> searchMovablePositions(Position target) {
-        List<Position> movablePositions = new ArrayList<>();
-
-        int horizontalWeight = target.getHorizontalWeight();
-        int verticalWeight = target.getVerticalWeight();
-
-        for (Direction direction : Direction.values()) {
-            int changedHorizontalWeight = horizontalWeight + direction.getX();
-            int changedVerticalWeight = verticalWeight + direction.getY();
-            movablePositions.addAll(calculateBoardPosition(changedHorizontalWeight, changedVerticalWeight));
-        }
-
-        return movablePositions;
-    }
-
-    private List<Position> calculateBoardPosition(int changedHorizontalWeight, int changedVerticalWeight) {
-        List<Position> result = new ArrayList<>();
-        if (isInBorder(changedHorizontalWeight, changedVerticalWeight)) {
-            result.add(Position.of(
-                    Horizontal.findFromWeight(changedHorizontalWeight),
-                    Vertical.findFromWeight(changedVerticalWeight))
-            );
-        }
-        return result;
-    }
-
-    private boolean isInBorder(int horizontalWeight, int verticalWeight) {
-        return horizontalWeight >= Board.MIN_BORDER && horizontalWeight <= Board.MAX_BORDER
-                && verticalWeight >= Board.MIN_BORDER && verticalWeight <= Board.MAX_BORDER;
+        KingMovement kingMovement = new KingMovement();
+        return kingMovement.searchMovablePositions(target);
     }
 }
