+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,12 +1,9 @@
 package chess.domain.piece;
 
-import chess.domain.board.Horizontal;
 import chess.domain.board.Position;
-import chess.domain.board.Vertical;
+import chess.domain.piece.moveStrategy.RookMovement;
 import chess.domain.piece.moveStrategy.ToEndOfLineStrategy;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public class Rook extends Piece {
@@ -19,19 +16,7 @@ public class Rook extends Piece {
 
     @Override
     public List<Position> searchMovablePositions(Position target) {
-        List<Position> movablePositions = new ArrayList<>();
-        Vertical targetVertical = target.getVertical();
-        Horizontal targetHorizontal = target.getHorizontal();
-
-        for (Horizontal horizontal : Horizontal.values()) {
-            movablePositions.add(Position.of(horizontal, targetVertical));
-        }
-        for (Vertical vertical : Vertical.values()) {
-            movablePositions.add(Position.of(targetHorizontal, vertical));
-        }
-
-        movablePositions.removeAll(Collections.singletonList(target));
-
-        return movablePositions;
+        RookMovement rookMovement = new RookMovement();
+        return rookMovement.searchMovablePositions(target);
     }
 }
