+++ src/main/java/chess/domain/piece/moveStrategy/MoveStrategy.java
@@ -1,0 +1,9 @@
+package chess.domain.piece.moveStrategy;
+
+import chess.domain.board.Position;
+
+import java.util.List;
+
+public interface MoveStrategy {
+    List<Position> searchMovablePositions(Position target);
+}
